refactor(Number.java) : 숫자 common 패키지 분리
