+++ docs/README.md
@@ -62,6 +62,10 @@
 # STEP3
 ## 로또(자동)
 ### 공통
+#### 숫자
+- [x] 로또 번호를 생성한다.
+  - [x] 45 초과의 수를 가질 수 없다.
+  - [x] 1 미만의 수를 가질 수 없다.
 ### 로또
 #### 로또 구입 금액
 - [x] 로또 구매 금액을 생성한다.
@@ -75,11 +79,7 @@
   - [x] 구매횟수만큼의 로또 번호 리스트를 생성한다.
   - [x] Collections.shuffle()을 활용한다.
 #### 로또 번호
-- [x] 로또 번호를 생성한다.
-  - [x] 45 초과의 수를 가질 수 없다.
-  - [x] 1 미만의 수를 가질 수 없다.
-#### 로또 번호들
-- [ ] 로또 번호를 추가한다.  
+- [ ] 숫자를 추가한다.  
   - [x] 6개 이상의 수를 추가 할 수 없다. 
   - [x] 중복된 수를 생성할 수 없다.
 #### 로또 번호 출력기
@@ -88,10 +88,8 @@
   - [ ] 로또 번호 구분은 , 마지막 로또 번호 뒤에는 찍지 않는다.
 ### 당첨
 #### 당첨 번호
-- [ ] 당첨 번호를 입력받는다.
+- [ ] 숫자를 추가한다.
   - [ ] 숫자만 입력 가능하다.
-  - [ ] 1 미만의 수를 가질 수 없다.
-  - [ ] 45 초과의 수를 가질 수 없다.
   - [ ] 중복된 수를 입력할 수 없다.
 - [x] 일치 갯수를 판별한다.
   - [x] 로또 번호들을 받아 일치 갯수를 판별한다.
