+++ src/main/java/common/vo/Number.java
@@ -1,4 +1,4 @@
-package lotto.lotto.domain;
+package common.vo;
 
 import java.util.Objects;
