+++ src/main/java/lotto/winning/domain/WinningNumber.java
@@ -1,6 +1,6 @@
 package lotto.winning.domain;
 
-import lotto.lotto.domain.Number;
+import common.vo.Number;
 import lotto.lotto.domain.Lotto;
 
 import java.util.ArrayList;
